Whoa there. A whole load of architecture
The AppDelegate now loads "Debug", which displays a list of all the
tests (this'll make it easy to develop and evaluate).

Test - class representing a test. holds the pre-test, actual test and post-test
PreTest and PostTest - For displaying information before and after tests

TestViewController - general class for the actual tests